Solved issue with deletets in testing
Reseting the application on the device was confusing sync services
because we were using the device's UUID as the client identifier.

This was an error all the way around and the sync client on the device
now generates and stores its own personal UUID.  This will allow
device recovery (which would reset the device UUID) and resolves
the issue with data deletion in testing.